+++ frontend/src/pages/Menu/styles.js
@@ -1,11 +1,6 @@
 import styled from 'styled-components';
-import React from 'react';
-import { BsFillPersonFill } from 'react-icons/bs';
-import { FiSettings, FiLogOut } from 'react-icons/fi';
-import { RiDashboardLine } from 'react-icons/ri';
-import PropTypes from 'prop-types';
 
-const MenuStyled = styled.nav`
+export const MenuStyled = styled.nav`
 	align-items: flex-start;
 	display: flex;
 	flex-direction: column;
@@ -18,7 +13,6 @@ const MenuStyled = styled.nav`
 
 	width: ${({ open }) => (open ? '280px' : '80px')};
 	background-color: ${(p) => p.theme.background};
-	//idk how to animate with flexbox transition: 0.1s;
 	border-right: 2px solid ${(p) => p.theme.textSecondary};
 
 	@media (max-width: ${({ theme }) => theme.mobile}) {
@@ -27,23 +21,18 @@ const MenuStyled = styled.nav`
 		background-color: ${(p) => p.theme.background};
 	}
 
-	a {
+	> a {
 		text-decoration: none;
 		font-size: 30px;
 		color: ${(p) => p.theme.text};
-		span {
+		> span {
 			visibility: ${({ open }) => (open ? 'visible' : 'hidden')};
-		}
-
-		i {
-			@media (max-width: ${({ theme }) => theme.mobile}) {
-				visibility: hidden;
-			}
 		}
 	}
 `;
 
-const Hamburger = styled.div`
+export const Hamburger = styled.div`
+	/* stylelint-disable -- idk. */
 	width: 30px;
 	height: 24px;
 	z-index: 10;
@@ -60,10 +49,6 @@ const Hamburger = styled.div`
 		background: ${(p) => (p.open ? p.theme.colors.error : p.theme.text)};
 	}
 
-	&:hover {
-		cursor: pointer;
-	}
-
 	& span:nth-child(1) {
 		transform: ${(p) => (p.open ? 'translateY(9px) rotate(45deg)' : '')};
 	}
@@ -78,60 +63,9 @@ const Hamburger = styled.div`
 		opacity: ${(p) => (p.open ? 0 : 1)};
 		transition: opacity 0.5s cubic-bezier(0.77, 0.2, 0.05, 1), background 0.5s;
 	}
+	/*stylelint-enable*/
 `;
 
-const Spacer = styled.div`
+export const Spacer = styled.div`
 	flex-grow: 1;
 `;
-
-const Menu = ({ open, setHamburger }) => {
-	return (
-		<MenuStyled open={open}>
-			<Hamburger
-				open={open}
-				onClick={() => {
-					setHamburger(!open);
-				}}
-			>
-				<span />
-				<span />
-				<span />
-			</Hamburger>
-
-			<a href="/#/">
-				<i>
-					<RiDashboardLine />
-				</i>
-				<span label="about us">Dashboard</span>
-			</a>
-			<a href="/#/profile">
-				<i>
-					<BsFillPersonFill />
-				</i>
-				<span label="about us">Profile</span>
-			</a>
-
-			<Spacer />
-			<a href="/#/settings">
-				<i>
-					<FiSettings />
-				</i>
-				<span label="settings">Settings</span>
-			</a>
-			<a href="/users/logout">
-				<i>
-					<FiLogOut />
-				</i>
-				<span label="logout">Logout</span>
-			</a>
-		</MenuStyled>
-	);
-};
-
-Menu.propTypes = {
-	open: PropTypes.bool,
-	setHamburger: PropTypes.func,
-};
-
-
-export default Menu;
